requirements.txt
django==3.0.8
pyodbc==4.0.30
django-jalali-date==0.3.1
